+++ pootle/apps/pootle_store/models.py
@@ -872,9 +872,10 @@ class Unit(models.Model, base.TranslationUnit):
         """Bring a unit's state back to life.
 
         :param is_fuzzy: whether the unit is in fuzzy state or not.
+        :return: `True` if the unit was resurrected, `False` otherwise.
         """
         if self.state > OBSOLETE:
-            return
+            return False
 
         if filter(None, self.target_f.strings):
             # when Unit toggles its OBSOLETE state the number of translated
@@ -890,6 +891,7 @@ class Unit(models.Model, base.TranslationUnit):
         self.update_qualitychecks(keep_false_positives=True)
         self._state_updated = True
         self._save_action = UNIT_RESURRECTED
+        return True
 
     def istranslated(self):
         if self._target_updated and not self.isfuzzy():